+++ Sources/RAW_hex/RAW_hex_enc_impl.swift
@@ -5,11 +5,11 @@ import RAW
 internal struct Encode {
 
 	/// computes the number of encoded bytes that would be required to encode the given number of unencoded bytes.
-	internal static func length(_ bytes:size_t) -> size_t {
+	internal static func length(_ bytes:Int) -> Int {
 		return bytes * 2
 	}
 
-	internal static func process_inline(decoded_data:UnsafePointer<UInt8>, encoded_index:size_t) -> (Value, Value) {
+	internal static func process_inline(decoded_data:UnsafePointer<UInt8>, encoded_index:Int) -> (Value, Value) {
 		let byte = decoded_data[encoded_index / 2]
 		let high = byte >> 4
 		let low = byte & 0x0F
